+++ gradle/libs.versions.toml
@@ -20,7 +20,7 @@ androidx-test-rules = "1.7.0"
 androidx-test-junit = "1.3.0"
 androidx-work = "2.11.2"
 bitfire-cert4android = "b8e2d33f09"
-bitfire-dav4jvm = "c7ba124427"
+bitfire-dav4jvm = "c06dafca54"
 compose-accompanist = "0.37.3"
 compose-bom = "2026.05.01"
 conscrypt = "2.5.3"
